+++ Assets/Game/Scripts/GDPR/Views/GDPRDlgView.cs
@@ -8,7 +8,6 @@ using strange.extensions.mediation.impl;
 using strange.extensions.signal.impl;
 using System;
 using TMPro;
-using HUFEXT.GenericGDPR.Runtime.Utils;
 using DG.Tweening;
 using System.Collections;
 
@@ -63,14 +62,14 @@ namespace TurboLabz.InstantFramework
             showRegularAdsBtn.onClick.AddListener(OnShowRegularAdsBtnClicked);
             acceptAndCollectBtn.onClick.AddListener(OnAcceptAndCollectBtnClicked);
 
-            noteText.text = "I hereby consent to the usage and disclosure of my personal data (including device information, advertising ID and my preferences) to​ <link=" +
-                GDPRTranslationsProvider.AD_PARTNERS_LINK + "><b><u>advertising network companies</u></b></link> for the purpose of serving tailored advertisements to me in the game.";
+            //noteText.text = "I hereby consent to the usage and disclosure of my personal data (including device information, advertising ID and my preferences) to​ <link=" +
+            //    GDPRTranslationsProvider.AD_PARTNERS_LINK + "><b><u>advertising network companies</u></b></link> for the purpose of serving tailored advertisements to me in the game.";
 
-            withdrawText.text = "I understand that I can withdraw this consent at any time within the game settings, as also described in our​ <link=" +
-                GDPRTranslationsProvider.PRIVACY_POLICY_LINK + "><b><u>Privacy Policy</u></b></link>.";
+            //withdrawText.text = "I understand that I can withdraw this consent at any time within the game settings, as also described in our​ <link=" +
+            //    GDPRTranslationsProvider.PRIVACY_POLICY_LINK + "><b><u>Privacy Policy</u></b></link>.";
 
-            noteText.gameObject.AddComponent<Hyperlink>();
-            withdrawText.gameObject.AddComponent<Hyperlink>();
+            //noteText.gameObject.AddComponent<Hyperlink>();
+            //withdrawText.gameObject.AddComponent<Hyperlink>();
         }
 
         public void OnServerDataAvailable()