+++ Assets/HUFEXT/PackageManager/Editor/Commands/Connection/GetPackageVersionsCommand.cs
@@ -18,10 +18,6 @@ namespace HUFEXT.PackageManager.Editor.Commands.Connection
                 Complete( false, "Unable to fetch versions from development channel." );
                 return;
             }
-
-            package.huf.config.stableVersions.Clear();
-            package.huf.config.previewVersions.Clear();
-            package.huf.config.previewVersions.Clear();
             EditorApplication.update += WaitForFinish;
             fetchesFinished = 0;
             fetchesCount = package.huf.scopes.Count * 3;
@@ -69,6 +65,7 @@ namespace HUFEXT.PackageManager.Editor.Commands.Connection
 
             void AddToVersionsList( List<Models.Version> configList, List<Models.Version> versions )
             {
+                configList.Clear();
                 foreach ( var version in versions )
                 {
                     if ( !configList.Exists( v => v.version == version.version ) )